requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
blinker==1.9.0
Brotli==1.0.9
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
comm==0.2.3
contourpy==1.3.0
cramjam==2.11.0
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastparquet==2024.11.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.59.1
freetype-py==2.3.0
fsspec==2025.7.0
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nest_asyncio==1.6.0
numpy==1.24.3
openai==1.6.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
patsy==1.0.1
pdfkit==1.0.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==25.3
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
protobuf==6.33.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.28.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pyparsing==3.2.3
pyphen==0.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.0.2
reportlab==4.4.1
requests==2.32.5
rlPyCairo==0.2.0
rpy2==3.5.11
scikit-learn==1.6.1
scipy==1.10.0
seaborn==0.13.2
setuptools==80.9.0
simplegeneric==0.8.1
six==1.17.0
sniffio==1.3.1
stack_data==0.6.3
statsmodels==0.14.5
tensorboardX==2.6.4
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3
unicodedata2==16.0.0
urllib3==2.5.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.4
wheel==0.45.1
wkhtmltopdf==0.2
zipp==3.23.0
zopfli==0.2.3.post1
zstandard==0.23.0
